// app/types/project.ts
export interface Project {
  id: number
  title: string
  description: string
  image: string
  technologies: string[]
  githubUrl: string
  category: string
}

export interface Video {
  id: number
  title: string
  description: string
  thumbnailUrl: string
  youtubeId: string
  category: string
}